+++ tests/test_intervention_count.py
@@ -1,4 +1,5 @@
 #  Copyright (C) 2025 by Higher Expectations for Racine County
+from polars import Int8
 
 from panoramel import InterventionCountContext
 
@@ -9,4 +10,4 @@ def test_el_intervention_count_context(mock_uuid):
     assert context.context_id == b"1"
     assert context.as_tuple() == (b"1", 42)
     assert context.output_name == "Number of Interventions"
-    assert context.output_type == int
+    assert context.output_type == Int8
